requirements.txt
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
passlib==1.7.4
PyJWT==2.3.0
Werkzeug==2.0.2

